backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg[binary]>=3.2.10
anthropic==0.40.0
azure-cognitiveservices-speech==1.41.1
python-dotenv==1.0.1
google-auth==2.29.0
whitenoise==6.8.2
